feature: add skeleton history
