+++ src/components/history/History.tsx
@@ -1,39 +1,17 @@
 import { Marker } from 'mapbox-gl';
-import { useLayoutEffect, useState } from 'react';
+import { useState } from 'react';
 import { useMap, usePlaces } from '../../hook';
 import { useHistory } from '../../hook/useHistory';
 import { IntSearch } from '../../interfaces/history.interface';
 import SearchSave from '../search-save/SearchSave';
 import './History.scss';
+import { HistorySkeleton } from '../ui/history-skeleton/HistorySkeleton';
 
 const History = () => {
-  const { history, setHistory } = useHistory();
+  const { history, isLoading } = useHistory();
   const { map, getRouteBetweenPoints } = useMap();
   const { userLocation } = usePlaces();
   const [activeMarker, setActiveMarker] = useState<Marker | null>(null);
-
-  useLayoutEffect(() => {
-    // Request for last five searches to backend
-    const url = 'http://localhost:8000/api/v1/addresses/recent';
-    const options = {
-      headers: {
-        'Content-Type': 'application/json',
-        'X-API-Key': `${import.meta.env.VITE_API_KEY_BACK}`,
-      },
-    };
-
-    fetch(url, options)
-      .then((res) => {
-        if (!res.ok) throw new Error('Error to obtain data');
-        return res.json();
-      })
-      .then(({ data }) => {
-        setHistory(data);
-      })
-      .catch((err) => {
-        console.error('Error to obtain data', err);
-      });
-  }, []);
 
   const handleClick = (place: IntSearch) => {
     const [lng, lat] = place.attributes.coordinates;
@@ -55,13 +33,15 @@ const History = () => {
   return (
     <article className="history">
       <h2 className="history__title">Most recent searches</h2>
-      <div className="history__container">
-        {history ? (
-          history.map((search) => <SearchSave key={search.id} search={search} handleClick={handleClick} />)
-        ) : (
-          <p className="history__advice">You don't have search history yet</p>
-        )}
-      </div>
+      {isLoading ? (
+        <HistorySkeleton />
+      ) : (
+        <div className="history__container">
+          {history.map((search) => (
+            <SearchSave key={search.id} search={search} handleClick={handleClick} />
+          ))}
+        </div>
+      )}
     </article>
   );
 };
