+++ src/components/ui/history-skeleton/HistorySkeleton.tsx
@@ -1,0 +1,13 @@
+
+import './HistorySkeleton.scss';
+
+export const HistorySkeleton = () => {
+  return (
+    <div className="skeleton">
+      <div className="skeleton__history"></div>
+      <div className="skeleton__history"></div>
+      <div className="skeleton__history"></div>
+      <div className="skeleton__history"></div>
+    </div>
+  );
+};
