+++ src/context/history/HistoryProvider.tsx
@@ -1,4 +1,4 @@
-import { useReducer } from 'react';
+import { useEffect, useReducer } from 'react';
 import { fetchAPIBackend, HistoryToSave } from '../../api';
 import { IntSearch } from '../../interfaces/history.interface';
 import { isInHistory } from '../../utils';
@@ -21,6 +21,28 @@ const INITIAL_STATE: HistoryState = {
 
 export const HistoryProvider = ({ children }: Props) => {
   const [state, dispatch] = useReducer(historyReducer, INITIAL_STATE);
+
+  useEffect(() => {
+    const url = 'http://localhost:8000/api/v1/addresses/recent';
+    const options = {
+      headers: {
+        'Content-Type': 'application/json',
+        'X-API-Key': `${import.meta.env.VITE_API_KEY_BACK}`,
+      },
+    };
+
+    fetch(url, options)
+      .then((res) => {
+        if (!res.ok) throw new Error('Error to obtain data');
+        return res.json();
+      })
+      .then(({ data }) => {
+        setHistory(data);
+      })
+      .catch((err) => {
+        console.error('Error to obtain data', err);
+      });
+  }, []);
 
   const setHistory = (history: IntSearch[]) => {
     dispatch({ type: 'SET_HISTORY', payload: history });
